feat: apply suspense to fetch card detail
