--- src/useGetData.js
@@ -1,48 +1,0 @@
-import { useState, useEffect } from 'react';
-import { defaultAxios } from 'apis/utils';
-
-const promiseWrapper = (promise) => {
-  let status = 'pending';
-  let result;
-
-  const s = promise.then(
-    (value) => {
-      status = 'success';
-      result = value;
-    },
-    (error) => {
-      status = 'error';
-      result = error;
-    },
-  );
-
-  return () => {
-    switch (status) {
-      case 'pending':
-        throw s;
-      case 'success':
-        return result;
-      case 'error':
-        throw result;
-      default:
-        throw new Error('Unknown status');
-    }
-  };
-};
-
-export function fetchCardList(url, config, deps) {
-  const [resource, setResource] = useState(null);
-
-  useEffect(() => {
-    const getData = async () => {
-      const promise = defaultAxios
-        .get(url, config)
-        .then((response) => response.data);
-      setResource(promiseWrapper(promise));
-    };
-
-    getData();
-  }, deps);
-
-  return resource;
-}
